[fix][plum] gihp\Object\Tree should re-evaluate subtrees on getSHA1()
Child trees can be updated without the parent noticing. This
causes stale SHA1s to be used for name mapping and in the objects.
It may also cause bad references to objects being written to disk.
So remap all names and objects when a new SHA1 is generated, and
clear the SHA1 immediately afterwards to always regenerate the SHA1